refactor: BoardUtility 분리
